Separating miscellaneous utilities into separate file
