+++ test/e2e/e2e.test.ts
@@ -1,14 +1,12 @@
 import { expect } from 'chai';
-import zlib from 'zlib';
 import path from 'path';
-import _ from 'lodash';
 import { getFiles, existsMustBeDir } from '../../src/utils';
+import { testNameToFilename, testNameToWorkingDirectory, afterRecord } from './e2e-utils';
 
 //Nock Config
-import { back as nockBack, Definition } from 'nock';
+import { back as nockBack } from 'nock';
 nockBack.setMode('record');
-nockBack.fixtures = __dirname + '/recordings'; //this only needs to be set once in your test helper
-type Recording = Definition & { rawHeaders: string[] }; // Hack to fix issue in Nock types that does not include `rawHeaders` property in `Definition` type
+nockBack.fixtures = __dirname + '/recordings';
 
 //Application Config
 import { dump, deploy } from '../../src';
@@ -97,50 +95,3 @@ describe('#end-to-end dump and deploy cycle', function () {
     nockDone();
   });
 });
-
-function decodeBuffer(recordingResponse: Definition['response']) {
-  try {
-    // Decode the hex buffer that nock made
-    const response = Array.isArray(recordingResponse)
-      ? recordingResponse.join('')
-      : recordingResponse;
-    const decoded = Buffer.from(response as string, 'hex');
-    const unzipped = zlib.gunzipSync(decoded).toString('utf-8');
-    return JSON.parse(unzipped);
-  } catch (err) {
-    throw new Error(`Error decoding nock hex:\n${err}`);
-  }
-}
-
-function afterRecord(recordings: Definition[]): Definition[] {
-  return recordings.map((recording) => {
-    recording.response = decodeBuffer(recording.response as string);
-    //@ts-ignore because we know `rawHeaders` actually exists
-    return sanitizeRecording(recording);
-  });
-}
-
-function testNameToFilename(testName = ''): string {
-  if (testName === '') throw new Error('Test name not defined');
-  return `${testName.replaceAll(' ', '-')}.json`;
-}
-
-function testNameToWorkingDirectory(testName = ''): string {
-  const directoryName = testName.replaceAll(' ', '-');
-  return path.join('./local/recorded', directoryName);
-}
-
-function sanitizeRecording(recording: Recording): Recording {
-  const sanitizedRecording = recording;
-
-  sanitizedRecording.rawHeaders = (() => {
-    const newHeaders = _.chunk(recording.rawHeaders, 2)
-      .filter((pair) => {
-        return pair[0] !== 'Content-Encoding'; // Prevents recordings from becoming gzipped
-      })
-      .flat();
-    return newHeaders;
-  })();
-
-  return sanitizedRecording;
-}
